pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retireplan"
version = "0.1.0"
description = "Deterministic retirement planning engine for veteran households"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Kevin", email = "kevin@example.com" }]
dependencies = ["pydantic>=2.0", "pyyaml>=6.0", "typer>=0.9.0", "PySide6>=6.7"]

[project.optional-dependencies]
dev = ["pytest>=7.4", "ruff>=0.1.0", "black>=23.0", "pre-commit>=3.0"]

[project.scripts]
retireplan = "retireplan.cli.main:app"

[tool.setuptools.packages.find]
include = ["retireplan*"]

[tool.setuptools.package-data]
retireplan = ["defaults/*.yaml", "defaults/*.csv"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # Pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
ignore = [
    "E501", # line too long (handled by black)
]

[tool.ruff.lint.isort]
known-first-party = ["retireplan"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
